Update ALLp2.py
The code is completely different from what I originally uploaded, therefore I have replaced it. The treasure scoring system does not work properly yet, I shall work on this tomorrow, also the background image is very important for this to work so download that and place both files in th same folder before running the code.